ZAPI-637: prevent VMAPI from exploding on GET /vms
Add input parameters validation for the ListVms endpoint (GET request
on /vms) and relevant tests.
Add tools/add-test-vms, initially used to reproduce the problem.

Add "marker" parameter to allow clients to paginate through entries
without the drawbacks of using "offset", that is seeing one or more
entries more than once if new entries are added while paginating.

The use of parameters "marker" and "offset" in the same request is not
allowed and results in a request error.

Remove the use of "noLimit = true" and force the limit of number of
entries returned to the maximum limit (currently 1000).

Allow the tests suite runner to run a subset of tests according to a
filter passed on the command line with "-f".