ZAPI-655: support "marker" for pagination on /vms
Add new parameter "marker" to handle pagination via VM markers instead
of offset.

Using "marker" prevents any duplicate entry to be listed if new entries
are added while paginating through a dataset. A marker is a JSON string
that uniquely identifies a VM object. See the changes made to the
documentation for more information.

The documentation is updated to deprecate usage of "offset" in favor of
"marker". It also documents that using both "offset" and "marker"
parameters is invalid and results in a request error.

This change also enforces a maximum limit of 1000 entries in the
response when listing VMs. This is likely to break existing clients that
do not use node-sdc-clients. As a consequence, this change also bumps
the version number in package.json to 8.0.0.

Finally, it also sorts by default on uuid descending, so that any
ListVms request can be used as the first page of a series of paginated
requests using marker={"uuid":"some-uuid"}.