ZAPI-668: fix offset_fields_vms test
Currently, the offset_fields_vms_ok test relies on more than two SDC
zones already being setup as 128MBs zones: assets0, dhcpd0 and other VMs
created during other test runs, or by other unrelated processes.

This change is a quick fix that only relies on assets0 and dhcpd0
existing and being setup as 128MBs zones. This is still not a proper
fix, but allows us to fix breakage now and work on a proper fix that
will not rely on any pre-existing zone.