DOCKER-557: fast `docker rm`
Do not wait until a VM is actually destroyed/deleted to return from a
DELETE /vms/vm_uuid request. Instead:

1. mark the VM as destroyed
2. stop the VM
3. at that point, do not allow any change to this VM except for marking
it as destroyed when it's actually been destroyed.

When a VM is in a 'destroyed' state, it cannot transition to any other
state, except to the state 'manual_override'. A VM can transition from
the 'manual_override' state to any other state.

This change also adds versioning to the current vmapi_vms bucket so that
it indexes the newly added 'zone_state' index required to be able to
search for VMs with `state === 'destroyed' && zone_state !== 'destroyed'`
when looking for "destroying" VMs.