ZAPI-685: do not start a workflow on DELETE /vms if not needed
When handling a DELETE /vms/some-vm request, do not start a workflow if
one of the following conditions are true:

1. The VM is in state 'provisioning'
2. The VM doesn't have a server_uuid
3. The server_uuid of the VM is not present in CNAPI

In cases 2 and 3, mark the VM as destroyed and return a 200 HTTP status
code. In case 1, return an error.

Returning early and not starting a destroy workflow in these cases is
faster and doesn't take any resource in the workflow's queue, thus
allowing other workflows who have a chance to succeed to be processed
more quickly.

PR: #14
PR-URL: https://github.com/joyent/sdc-vmapi/pull/14
Reviewed-By: Josh Wilsdon <jwilsdon@joyent.com>
Reviewed-By: Richard Kiene <richard.kiene@joyent.com>